chore: move Viewer index component to index file
